+++ src/map/node.rs
@@ -220,9 +220,9 @@ impl<T> Node<T> {
     arena: &'a Arena,
     trailer: T,
     value_size: u32,
-    f: impl FnOnce(&mut VacantValue<'a>) -> Result<(), E>,
+    f: impl FnOnce(&mut VacantBuffer<'a>) -> Result<(), E>,
   ) -> Result<(), Either<E, Error>> {
-    let offset = Self::new_value(arena, trailer, value_size, f)?;
+    let (offset, value_size) = Self::new_value(arena, trailer, value_size, f)?;
     let (_, old_size) = self.value.swap(offset, value_size);
 
     // on success, which means that old value is removed, we need to incr the discard bytes
@@ -268,13 +268,13 @@ impl<T> Node<T> {
     arena: &'a Arena,
     trailer: T,
     value_size: u32,
-    f: impl FnOnce(&mut VacantValue<'a>) -> Result<(), E>,
-  ) -> Result<u32, Either<E, Error>> {
+    f: impl FnOnce(&mut VacantBuffer<'a>) -> Result<(), E>,
+  ) -> Result<(u32, u32), Either<E, Error>> {
     if value_size as u64 > u32::MAX as u64 {
       return Err(Either::Right(Error::ValueTooLarge(value_size as u64)));
     }
 
-    let value_offset = arena
+    let (value_offset, allocated) = arena
       .alloc_value::<T>(value_size)
       .map_err(|e: ArenaError| Either::Right(e.into()))?;
 
@@ -286,8 +286,12 @@ impl<T> Node<T> {
       ptr::write(trailer_ptr, trailer);
 
       let val = core::slice::from_raw_parts_mut(ptr.add(mem::size_of::<T>()), value_size as usize);
-      Self::fill_vacant_value(arena, value_size, val, f).map_err(Either::Left)?;
-      Ok(value_offset)
+      let value_size =
+        Self::fill_vacant_value(arena, value_size, value_offset, val, f).map_err(|e| {
+          arena.incr_discard(allocated);
+          Either::Left(e)
+        })?;
+      Ok((value_offset, value_size as u32))
     }
   }
 
@@ -295,10 +299,11 @@ impl<T> Node<T> {
   unsafe fn fill_vacant_value<'a, E>(
     arena: &'a Arena,
     value_size: u32,
+    value_offset: u32,
     buf: &'a mut [u8],
-    f: impl FnOnce(&mut VacantValue<'a>) -> Result<(), E>,
-  ) -> Result<(), E> {
-    let mut oval = VacantValue::new(value_size as usize, buf);
+    f: impl FnOnce(&mut VacantBuffer<'a>) -> Result<(), E>,
+  ) -> Result<u32, E> {
+    let mut oval = VacantBuffer::new(value_size as usize, value_offset, buf);
     f(&mut oval)?;
     let remaining = oval.remaining();
     if remaining != 0 {
@@ -307,7 +312,7 @@ impl<T> Node<T> {
 
       arena.incr_discard(remaining as u32);
     }
-    Ok(())
+    Ok(oval.len() as u32)
   }
 }
 
@@ -318,7 +323,7 @@ impl<T: Trailer> Node<T> {
     key: &'b [u8],
     trailer: T,
     value_size: u32,
-    f: impl FnOnce(&mut VacantValue<'a>) -> Result<(), E>,
+    f: impl FnOnce(&mut VacantBuffer<'a>) -> Result<(), E>,
   ) -> Result<NodePtr<T>, Either<E, Error>> {
     if height < 1 || height > MAX_HEIGHT as u32 {
       panic!("height cannot be less than one or greater than the max height");
@@ -346,7 +351,7 @@ impl<T: Trailer> Node<T> {
     let AllocMeta {
       node_offset,
       value_offset,
-      allocated: _,
+      allocated,
     } = arena
       .alloc::<T>(
         node_size + key_size as u32,
@@ -375,8 +380,88 @@ impl<T: Trailer> Node<T> {
       #[cfg(not(feature = "unaligned"))]
       ptr::write(trailer_ptr, trailer);
 
-      Self::fill_vacant_value(arena, value_size, node.get_value_mut(arena), f)
-        .map_err(Either::Left)?;
+      let value_size = Self::fill_vacant_value(
+        arena,
+        value_size,
+        value_offset,
+        node.get_value_mut(arena),
+        f,
+      )
+      .map_err(|e| {
+        arena.incr_discard(allocated);
+        Either::Left(e)
+      })?;
+
+      node.value = Pointer::new(value_offset, value_size);
+      Ok(NodePtr::new(ptr, node_offset))
+    }
+  }
+
+  pub(super) fn new_node_ptr_with_key<'a, 'b: 'a, E>(
+    arena: &'a Arena,
+    height: u32,
+    key_offset: u32,
+    key_size: u16,
+    trailer: T,
+    value_size: u32,
+    f: impl FnOnce(&mut VacantBuffer<'a>) -> Result<(), E>,
+  ) -> Result<NodePtr<T>, Either<E, Error>> {
+    if height < 1 || height > MAX_HEIGHT as u32 {
+      panic!("height cannot be less than one or greater than the max height");
+    }
+
+    if value_size as u64 > u32::MAX as u64 {
+      return Err(Either::Right(Error::ValueTooLarge(value_size as u64)));
+    }
+
+    let entry_size = (value_size as u64) + (key_size as u64) + Node::<T>::MAX_NODE_SIZE;
+    if entry_size > u32::MAX as u64 {
+      return Err(Either::Right(Error::EntryTooLarge(entry_size)));
+    }
+
+    // Compute the amount of the tower that will never be used, since the height
+    // is less than maxHeight.
+    let unused_size = (MAX_HEIGHT as u32 - height) * (Link::SIZE as u32);
+    let node_size = (Self::MAX_NODE_SIZE as u32) - unused_size;
+
+    let AllocMeta {
+      node_offset,
+      value_offset,
+      allocated,
+    } = arena
+      .alloc::<T>(node_size, value_size, Self::ALIGN, unused_size)
+      .map_err(|e| Either::Right(e.into()))?;
+
+    unsafe {
+      // Safety: we have check the offset is valid
+      let ptr = arena.get_pointer_mut(node_offset as usize);
+      // Safety: the node is well aligned
+      let node = &mut *(ptr as *mut Node<T>);
+      node.value = Pointer::new(value_offset, value_size);
+      node.key_offset = key_offset;
+      node.key_size = key_size;
+      node.height = height as u16;
+
+      #[cfg(not(feature = "unaligned"))]
+      ptr::write_bytes(ptr.add(mem::size_of::<Node<T>>()), 0, height as usize);
+
+      let ptr = arena.get_pointer_mut(value_offset as usize);
+      let trailer_ptr = ptr as *mut T;
+      #[cfg(not(feature = "unaligned"))]
+      ptr::write(trailer_ptr, trailer);
+
+      let value_size = Self::fill_vacant_value(
+        arena,
+        value_size,
+        value_offset,
+        node.get_value_mut(arena),
+        f,
+      )
+      .map_err(|e| {
+        arena.incr_discard(allocated);
+        Either::Left(e)
+      })?;
+      node.value = Pointer::new(value_offset, value_size);
 
       Ok(NodePtr::new(ptr, node_offset))
     }